Spring-Boot_Junit/src/main/java/com/stg/service/EmployeeService.java
package com.stg.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.stg.entity.Employee;
import com.stg.repository.EmployeeRepository;

import java.util.List;
import java.util.Optional;

@Service
public class EmployeeService {
    @Autowired
    private EmployeeRepository employeeRepository;

    public Employee addEmployee(Employee employee) {
        return employeeRepository.save(employee);
    }

    public Optional<Employee> getEmployee(int empId) {
        return employeeRepository.findById(empId);
    }

    public Employee updateEmployee(Employee employee) {
        return employeeRepository.save(employee);
    }

    public void deleteEmployee(int empId) {
        employeeRepository.deleteById(empId);
    }

    public List<Employee> listAllEmployees() {
        return employeeRepository.findAll();
    }
}